[VFS/STFS] Update L1/L2 free-block counts when allocating/freeing
TODO list:

- Hash tables are "created" by STFSGetHashTable, the fread() for it will fail which leaves the table empty-initialized, and that empty table is then used as the new table.
This works fine for creating new L0 hash blocks, but if the package is expanding and L1/L2 blocks have to be created, this won't handle populating them with hashes of existing hash blocks yet. Probably a simple way to handle that but not sure what it is atm.

Likely won't be a problem in 99% of cases, but if eg. a package had exactly 170 blocks already (so only needs a single L0 table), adding a new block will create a new L0 table, requiring an L1 table to be created to hash the two L0 tables, the new block will be marked as dirty, so the L1 entry for the new L0 table will be recalced by STFSFlush, but since none of the older 170 blocks were marked dirty, the hash of their L0 table won't be written into L1.
Need to find some way to keep track of newly added L1/L2 blocks, and then recalc them, or detect when STFSGetHashTable creates new L1/L2 tables and populate it there. Not sure what's most efficient way of handling it yet.

- Deleting a file & freeing the blocks for it only marks the blocks as free, the contents aren't deleted & package won't be resized - only way to resize file would be moving the free block to the end somehow. Could at least add code to resize file if the last block is freed I guess.

- ContentPackage class creates the StfsContainerDevice, but then std::move's it to use with VirtualFileSystem::RegisterDevice, I'm not all that sure about C++ smart pointers, but is there some way ContentPackage can retain a pointer to the device safely? or can we query VirtualFileSystem with the device path to get a pointer to it? (this is needed for getting/setting thumbnail on opened packages...)

- Secondary hash blocks aren't really handled by the writing side of the code, instead if we need to read in a secondary block (to get the correct hash entry data etc), the active_index flag will be unset after reading it, and the block will be treated as if we'd read it from the primary block - then if any changes are made, the updated active_index flag will get written, and the secondary-block we read in will overwrite the primary block instead.

I'm not sure how the X360 itself handles this, I guess it swaps between them after each write or something, would be nice if we could eventually support the "transactional" nature of it.