[Kernel] Check XEXP applicability via patch header data instead of exec-info
Some TUs can contain XEXPs for multiple discs, where each disc uses a different media-ID. The TU package header would only contain executable-info/media-ID for the first disc in this case.

Unfortunately we can't just check with the XEXP's execution-info as not every XEXP contains it for some reason, fortunately every XEXP does contain a digest of the target XEX's signature area though, so we can use that to tell if the XEXP is applicable or not.