[STFS] Fix crash when device is closed before the files themselves
BlazBlue seems to close the files after the device is closed, so when files try to access device during the destruction process it causes problems.
Fixed this by updating the StfsContainerEntry to know when device has been removed - not sure if this is the best way to handle this though.